Make the error initializer and instance initializer
This allows us to use the ajax service so errors written to the API
will be authenticated.  Also removes one of our deprecation notices.
